Proba

+++ ArkTools3.0/ArkTools3.0/Starter.cs
@@ -40,38 +40,42 @@ namespace ArkTools3._0
         private void BTNCreate_Click(object sender, EventArgs e)
         {
             string alma = "[Startup]" + Environment.NewLine;
-            string[] beolvas = File.ReadAllLines("settings.szomoroggyameg");
-            if (BTNFaLomb.Checked == true)
+            if (File.Exists("settings.szomoroggyameg"))
             {
-                alma += "foliage.UseOcclusionType=0" + Environment.NewLine;
+                string[] beolvas = File.ReadAllLines("settings.szomoroggyameg");
+                if (BTNFaLomb.Checked == true)
+                {
+                    alma += "foliage.UseOcclusionType=0" + Environment.NewLine;
+                }
+                if (BTNFaTorzs.Checked == true)
+                {
+                    alma += "foliage.CullAll=1" + Environment.NewLine;
+                }
+                alma += "ShowFlag.Fog=1" + Environment.NewLine;
+                if (BTNPlayerEffect.Checked == true)
+                {
+                    alma += "ShowFlag.Materials=0" + Environment.NewLine;
+                }
+                if (BTNArnyek.Checked == true)
+                {
+                    alma += "ShowFlag.LevelColoration=0" + Environment.NewLine + "ShowFlag.LightComplexity=0" + Environment.NewLine + "ShowFlag.LightInfluences=0" + Environment.NewLine + "ShowFlag.LightMapDensity=0" + Environment.NewLine + "ShowFlag.LightRadius=0" + Environment.NewLine + "ShowFlag.LightShafts=0" + Environment.NewLine + "ShowFlag.PreviewShadowsIndicator=0" + Environment.NewLine;
+                }
+                if (BTNNonPlayerEffect.Checked == true)
+                {
+                    alma += "ShowFlag.GameplayDebug=0" + Environment.NewLine;
+                }
+                if (BTNKod.Checked == true)
+                {
+                    alma += "ShowFlag.DynamicShadows=1" + Environment.NewLine;
+                }
+                if (BTNMaxFPS.Checked == true)
+                {
+                    alma += "t.maxfps=" + NUMInputMaxFps.Value + Environment.NewLine;
+                }
+                string path = beolvas[1] + @"\Engine\Config\ConsoleVariables.ini";
+                File.WriteAllText(path, alma);
             }
-            if (BTNFaTorzs.Checked == true)
-            {
-                alma += "foliage.CullAll=1" + Environment.NewLine;
-            }
-            alma += "ShowFlag.Fog=1" + Environment.NewLine;
-            if (BTNPlayerEffect.Checked == true)
-            {
-                alma += "ShowFlag.Materials=0" + Environment.NewLine;
-            }
-            if (BTNArnyek.Checked == true)
-            {
-                alma += "ShowFlag.LevelColoration=0" + Environment.NewLine + "ShowFlag.LightComplexity=0" + Environment.NewLine + "ShowFlag.LightInfluences=0" + Environment.NewLine + "ShowFlag.LightMapDensity=0" + Environment.NewLine + "ShowFlag.LightRadius=0" + Environment.NewLine + "ShowFlag.LightShafts=0" + Environment.NewLine + "ShowFlag.PreviewShadowsIndicator=0" + Environment.NewLine;
-            }
-            if (BTNNonPlayerEffect.Checked == true)
-            {
-                alma += "ShowFlag.GameplayDebug=0" + Environment.NewLine;
-            }
-            if (BTNKod.Checked == true)
-            {
-                alma += "ShowFlag.DynamicShadows=1" + Environment.NewLine;
-            }
-            if (BTNMaxFPS.Checked == true)
-            {
-                alma += "t.maxfps=" + NUMInputMaxFps.Value + Environment.NewLine;
-            }
-            string path = beolvas[1] + @"\Engine\Config\ConsoleVariables.ini";
-            File.WriteAllText(path, alma);
+            
             
         }
 
@@ -83,17 +87,20 @@ namespace ArkTools3._0
 
         private void siticoneButton6_Click(object sender, EventArgs e)
         {
-            string[] beolvas = File.ReadAllLines("settings.szomoroggyameg");
-            string path = beolvas[1] + @"\ShooterGame\Saved\LocalProfiles\PlayerLocalData.arkprofile";
-            if (File.Exists(path))
+            if (File.Exists("settings.szomoroggyameg"))
             {
-                File.Delete(path);
-                File.Copy("PlayerLocalData.arkprofile", path);
+                string[] beolvas = File.ReadAllLines("settings.szomoroggyameg");
+                string path = beolvas[1] + @"\ShooterGame\Saved\LocalProfiles\PlayerLocalData.arkprofile";
+                if (File.Exists(path))
+                {
+                    File.Delete(path);
+                    File.Copy("PlayerLocalData.arkprofile", path);
 
-            }
-            else
-            {
-                File.Copy("PlayerLocalData.arkprofile", path);
+                }
+                else
+                {
+                    File.Copy("PlayerLocalData.arkprofile", path);
+                }
             }
             
         }
@@ -128,6 +135,18 @@ namespace ArkTools3._0
             }
             MessageBox.Show("Done! Use -culture=global at steam!");
 
+        }
+
+        private void siticoneButton8_Click(object sender, EventArgs e)
+        {
+            if (File.Exists("Chars/RobotoDistanceField.uasset"))
+            {
+                MessageBox.Show("Létezik te pahaszt!");
+            }
+            else
+            {
+                MessageBox.Show("A francot képzelsz magadról!");
+            }
         }
     }
 }
